Upgrade zookeeper samples
