+++ 3-extensions/registry/dubbo-samples-zookeeper/case-versions.conf
@@ -20,7 +20,7 @@
 # Supported component versions of the test case
 
 # Spring app
-dubbo.version=2.7*, 3.*
-spring.version=4.*, 5.*
-java.version= [>= 8]
+dubbo.version=3.3.*
+spring.version=6.*
+java.version= [ >= 17]
 zookeeper-server.version=["!*-temurin", >= 3.6.0 ]
