+++ 3-extensions/registry/dubbo-samples-zookeeper/dubbo-samples-zookeeper-consumer/src/main/java/org/apache/dubbo/samples/ConsumerApplication.java
@@ -1,0 +1,49 @@
+/*
+ *
+ *   Licensed to the Apache Software Foundation (ASF) under one or more
+ *   contributor license agreements.  See the NOTICE file distributed with
+ *   this work for additional information regarding copyright ownership.
+ *   The ASF licenses this file to You under the Apache License, Version 2.0
+ *   (the "License"); you may not use this file except in compliance with
+ *   the License.  You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ *
+ */
+
+package org.apache.dubbo.samples;
+
+import org.apache.dubbo.config.spring.context.annotation.EnableDubbo;
+import org.apache.dubbo.samples.action.GreetingServiceConsumer;
+import org.springframework.beans.BeansException;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.ApplicationContext;
+import org.springframework.context.ApplicationContextAware;
+
+@SpringBootApplication
+@EnableDubbo
+public class ConsumerApplication implements ApplicationContextAware {
+
+    private static ApplicationContext context;
+
+    public static void main(String[] args) {
+        SpringApplication.run(ConsumerApplication.class, args);
+        GreetingServiceConsumer consumer = (GreetingServiceConsumer)context.getBean("annotatedConsumer");
+        System.out.println(consumer.doSayHello("zookeeper"));
+
+        System.exit(0);
+    }
+
+    @Override
+    public void setApplicationContext(ApplicationContext applicationContext) throws BeansException {
+        context = applicationContext;
+    }
+
+}
