+++ 3-extensions/registry/dubbo-samples-zookeeper/dubbo-samples-zookeeper-consumer/src/test/java/org/apache/dubbo/samples/GreetingServiceIT.java
@@ -1,0 +1,43 @@
+/*
+ *   Licensed to the Apache Software Foundation (ASF) under one or more
+ *   contributor license agreements.  See the NOTICE file distributed with
+ *   this work for additional information regarding copyright ownership.
+ *   The ASF licenses this file to You under the Apache License, Version 2.0
+ *   (the "License"); you may not use this file except in compliance with
+ *   the License.  You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *   Unless required by applicable law or agreed to in writing, software
+ *   distributed under the License is distributed on an "AS IS" BASIS,
+ *   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *   See the License for the specific language governing permissions and
+ *   limitations under the License.
+ */
+package org.apache.dubbo.samples;
+
+import org.apache.dubbo.samples.ConsumerApplication;
+import org.apache.dubbo.samples.action.GreetingServiceConsumer;
+import org.junit.Assert;
+import org.junit.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
+import org.springframework.test.context.junit4.SpringRunner;
+
+
+@SpringBootTest(classes = {ConsumerApplication.class})
+@RunWith(SpringRunner.class)
+@ExtendWith(SpringExtension.class)
+public class GreetingServiceIT {
+
+    @Autowired
+    private GreetingServiceConsumer consumer;
+
+    @Test
+    public void test() {
+        Assert.assertEquals("hello, zookeeper", consumer.doSayHello("zookeeper"));
+    }
+}
